pyproject.toml
[build-system]
requires = ["maturin>=1.8,<2.0"]
build-backend = "maturin"

[project]
name = "rxing"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dynamic = ["version"]
dependencies = [
    "Pillow",
    "numpy",
]

[tool.maturin]
features = ["pyo3/extension-module"]
python-source = "python"
module-name = "rxing.rxing_lib"
